+++ src/backend/joanie/core/filters/admin/__init__.py
@@ -15,9 +15,9 @@ from joanie.core import enums, models
 from joanie.core.filters.base import MultipleValueFilter
 from joanie.core.utils import get_default_currency_symbol
 
-from .enrollment import EnrollmentAdminFilterSet
-from .skill import SkillAdminFilterSet
-from .teacher import TeacherAdminFilterSet
+from .enrollment import EnrollmentAdminFilterSet  # pylint: disable=unused-import
+from .skill import SkillAdminFilterSet  # pylint: disable=unused-import
+from .teacher import TeacherAdminFilterSet  # pylint: disable=unused-import
 
 
 class OrganizationAdminFilterSet(filters.FilterSet):
@@ -348,14 +348,16 @@ class DiscountAdminFilterSet(filters.FilterSet):
             return Q(amount__icontains=amount)
 
         def rate_filter(rate, exact=False):
+            exact_rate = rate
             try:
-                rate = int(rate) / 100
+                exact_rate = int(exact_rate) / 100
             except ValueError:
                 pass
 
-            if exact and rate:
-                return Q(rate=rate)
-            return Q(rate__icontains=rate)
+            if exact and exact_rate:
+                return Q(rate=exact_rate)
+
+            return Q(rate__icontains=rate) | Q(rate__icontains=exact_rate)
 
         currency_symbol = get_default_currency_symbol()
         if currency_symbol in value: