+++ src/backend/joanie/core/flows/__init__.py
@@ -1,3 +1,4 @@
+# pylint: disable=unused-import
 """
 Viewflow Flows of Joanie "Core" application
 https://github.com/viewflow/viewflow