+++ src/backend/joanie/tests/base.py
@@ -70,6 +70,16 @@ class BaseAPITestCase(PerfRecTestCaseMixin, TestCase):
         """
         return generate_jwt_token_from_user(user, expires_at)
 
+    def assertStatusCodeEqual(self, response, status_code):
+        """
+        Assert that the response status code is equal to the expected status code.
+        """
+        try:
+            response_content = response.json()
+        except ValueError:
+            response_content = response.content
+        self.assertEqual(response.status_code, status_code, response_content)
+
 
 class LoggingTestCase(PerfRecTestCaseMixin, TestCase):
     """Base test case for logging tests"""